(refactor) Inline UDP
